requirements.txt
fastapi~=0.110.0
redis
uvicorn
langchain~=0.1.11
openai
tiktoken
urllib3
bs4
selenium
unstructured
python-decouple
nltk
python-magic
python-dotenv
pydantic~=2.6.3
transformers
langchain-community
typing
mysql-connector-python~=8.3.0
starlette~=0.36.3
python-docx
pypd
langchain-openai
yt_dlp
pydub
librosa